# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Article API",
    "description": "Article API to manage Admin users, Tenants, Articles, Tags and References.",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/api/",
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/tags": {
      "post": {
        "tags": ["Tag"],
        "description": "Create new Tag",
        "operationId": "createTag",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Tag Parameters",
            "in": "body",
            "description": "Create new tag",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "403": {
            "description": "Tag Duplication",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },      
      "get": {
        "tags": ["Tag"],
        "description": "Get Tag List",
        "operationId": "getTags",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Tag List",
            "schema": {
              "$ref": "#/definitions/GetTag"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/tags/{_id}": {
      "patch": {
        "tags": ["Tag"],
        "description": "Update Tag",
        "operationId": "updateTag",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Tag Parameters",
            "in": "body",
            "description": "Update Tag",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateTag"
            }
          },
          {
            "name": "_id",
            "in": "path",
            "description": "Tag Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Invalid Tag Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": ["Tag"],
        "description": "Delete Tag",
        "operationId": "deleteTag",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          },
          {
            "name": "_id",
            "in": "path",
            "description": "_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "404": {
            "description": "Invalid Tag Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/languages": {
      "post": {
        "tags": ["Language"],
        "description": "Create new Language",
        "operationId": "createLanguage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Language Parameters",
            "in": "body",
            "description": "Create new language",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "403": {
            "description": "Language Duplication",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },      
      "get": {
        "tags": ["Language"],
        "description": "Get Language List",
        "operationId": "getLanguages",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Language List",
            "schema": {
              "$ref": "#/definitions/GetLanguage"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/languages/{_id}": {
      "patch": {
        "tags": ["Language"],
        "description": "Update Language",
        "operationId": "updateLanguage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Language Parameters",
            "in": "body",
            "description": "Update Language",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateLanguage"
            }
          },
          {
            "name": "_id",
            "in": "path",
            "description": "Language Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Invalid Language Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": ["Language"],
        "description": "Delete Language",
        "operationId": "deleteLanguage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          },
          {
            "name": "_id",
            "in": "path",
            "description": "_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "404": {
            "description": "Invalid Language Id",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/subscriptions": {
      "patch": {
        "tags": ["Subscription"],
        "description": "Update Subcription",
        "operationId": "updateSubscription",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "parameter",
            "in": "body",
            "description": "parameter",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Subscription"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/subscriptions/{articleId}": {
      "get": {
        "tags": ["Subscription"],
        "description": "Get Subcription",
        "operationId": "getSubscription",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          },
          {
            "name": "articleId",
            "in": "path",
            "description": "Article Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "$ref": "#/definitions/GetSubscription"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "UpdateTag": {
      "properties": {
        "tag": {
          "required":true,
          "type": "string",
          "example": "Tag"
        },
        "token": {
          "required":true,
          "type": "string",
          "example": "Token"
        }
      }
    },
    "GetTag": {
      "properties": {
        "tag": {
          "required":true,
          "type": "string",
          "example": "Tag"
        }
      }
    },
    "Tag": {
      "properties": {
        "tag": {
          "required":true,
          "type": "string",
          "example": "Tag"
        },
        "token": {
          "required":true,
          "type":"string",
          "example": "Token"
        } 
      }
    },
    "UpdateLanguage": {
      "properties": {
        "language": {
          "required":true,
          "type": "string",
          "example": "language"
        },
        "token": {
          "required":true,
          "type": "string",
          "example": "Token"
        }
      }
    },
    "GetLanguage": {
      "properties": {
        "language": {
          "required":true,
          "type": "string",
          "example": "language"
        }
      }
    },
    "Language": {
      "properties": {
        "language": {
          "required":true,
          "type": "string",
          "example": "language"
        },
        "token": {
          "required":true,
          "type":"string",
          "example": "Token"
        } 
      }
    },
...
    "Subscription": {
      "properties": {
        "articleId": {
          "required":true,
          "type": "string",
          "example": "Article Id"
        },
        "subscription": {
          "required":true,
          "type": "boolean",
          "example": true
        },
        "token": {
          "required":true,
          "type": "string",
          "example": "Token"
        }
      }
    },
    "GetSubscription": {
      "properties": {
        "result": {
          "required":true,
          "type": "boolean"
        }
      }
    }
  }
}